erro no get favorite game

+++ game-service/src/main/java/com/TP/game_service/models/DTOs/FavoriteGameRequestDTO.java
@@ -3,5 +3,5 @@ package com.TP.game_service.models.DTOs;
 import java.util.UUID;
 
 public record FavoriteGameRequestDTO (
-        int gameId
+        Long gameId
 ) { }
